technical debt

+++ models/geo_location.go
@@ -19,8 +19,8 @@
 package models
 
 type Point struct {
-	Lon float64 `json:"lon"`
-	Lat float64 `json:"lat"`
+	Lon float32 `json:"lon"`
+	Lat float32 `json:"lat"`
 }
 
 // "status": "success",
